initial commit - delta batch import deals
